apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
opencv-python
numpy
pillow
pymupdf
reportlab
tqdm
